requirements.txt
# step one requirements
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Step 2: transcription + embeddings
faster-whisper==1.0.3
sentence-transformers==3.0.1
numpy==1.26.4
